requirements.txt
annotated-types==0.7.0
attrs==23.2.0
boto3==1.35.19
botocore==1.35.19
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
cloudpickle==2.2.1
dill==0.3.8
docker==7.1.0
filelock==3.16.1
fsspec==2024.9.0
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
googleapis-common-protos==1.65.0
httplib2==0.22.0
huggingface-hub==0.25.0
idna==3.8
importlib-metadata==6.11.0
iniconfig==2.0.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
logging==0.4.9.6
markdown-it-py==3.0.0
mdurl==0.1.2
mock==4.0.3
multiprocess==0.70.16
numpy==1.26.4
oauthlib==3.2.2
packaging==24.1
pandas==2.2.3
pathos==0.3.2
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
pox==0.3.4
ppft==1.7.6.8
proto-plus==1.24.0
protobuf==4.25.5
psutil==6.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.4
pytest==8.3.3
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
requests_auth==8.0.0
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
sagemaker==2.232.1
sagemaker-core==1.0.6
schema==0.7.7
setuptools==75.1.0
six==1.16.0
smdebug-rulesconfig==1.0.1
tblib==3.0.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
uuid==1.30
zipp==3.20.2
